Renamed patterns -> demographic_patterns
