+++ core_data_modules/cleaners/swahili/demographic_patterns.py
